comfoserver - use range 1-4 for addressing fan profiles globally
As documented in the handler itself, SetFanProfile() now takes a
profile identifier that ranges between 1-4 instead of the previous 0-3.

This was confusing because profiles are only 0-indexed when issuing
set commands, but 1-indexed everywhere else. This change makes for a
less-confusing user experience, the API now uses 1-4 everywhere when
addressing profiles.